tools/old_discord_method/requirements.txt
discord.py==2.1.0
Pillow==9.3.0
python-dotenv==0.19.0